Cleanup: fix potential overflow, remove dead code, unify angle normalization
- line2Dup.cpp: defensive int64 cast for magnitude squared (prevents
  overflow with non-standard Sobel kernels)
- roi_refine.cpp: guard empty constraints vector before median access (UB fix),
  reuse sorted distances in outlier filter, use std::move for filtered vector
- roi_refine.cpp: remove 160-line validateConstraints_OLD (#if 0 block)
- shape_matcher.cpp: remove 165-line evaluateQuality_OLD (#if 0 block)
- icp_refine.cpp, roi_refine.cpp, shape_matcher.cpp: replace while-loop angle
  normalization with std::fmod (7 locations)

+++ line2Dup.cpp
@@ -461,7 +461,9 @@ static void quantizedOrientations(const Mat &src, Mat &magnitude,
             // Scalar tail
             for (; c < src.cols - 1; ++c) {
                 int gx = dx[c], gy = dy[c];
-                int mag_sq_i = gx*gx + gy*gy;
+                // Use int64 to avoid overflow: max Sobel 3x3 on uint8 is 1020,
+                // so 1020^2+1020^2=2M fits int32, but defensive for other kernels.
+                int mag_sq_i = (int)((int64_t)gx*gx + (int64_t)gy*gy);
                 mag_r[c] = (float)mag_sq_i;
 
                 if (mag_sq_i <= threshold_sq_i) continue;
